+++ src/main/java/com/game/void_seekers/tools/Coordinates.java
@@ -1,14 +1,14 @@
 package com.game.void_seekers.tools;
 
 public class Coordinates {
-    public double x;
-    public double y;
+    public int x;
+    public int y;
 
     public Coordinates() {
         this(0, 0);
     }
 
-    public Coordinates(double x, double y) {
+    public Coordinates(int x, int y) {
         setX(x);
         setY(y);
     }
@@ -17,19 +17,19 @@ public class Coordinates {
         return Math.sqrt(Math.pow((x - other.x), 2) + Math.pow((y - other.y), 2));
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 }
